Move the header component to common.
