+++ src/components/common/header/Header.scss
@@ -1,4 +1,4 @@
-@import '../../../_variables.scss';
+@import '../../../variables';
 
 .review-header {
   display: flex;
